+++ fe2o3-amqp-ws/src/wasm.rs
@@ -1,3 +1,5 @@
+//! WebSocket in wasm with web-sys
+
 use web_sys::{WebSocket, MessageEvent};
 
 struct WasmWebSocketStream(WebSocket);
